requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-cors==3.0.10
flask-login==0.6.3
werkzeug==2.3.7

tensorflow==2.16.1
numpy==1.26.4
pillow==10.2.0

python-dotenv==1.0.1
